Complete django-double-entry extraction from lukweb repo
This preserves most of the history behind core components of the
accounting system, but of course an operation like this can never hope
to keep everything intact. In particular, the commits in this rewritten
history no longer represent working code, but should help somewhat in
keeping track of how things changed without having to go back to the
lukweb repo.